requirements.txt
numpy
pandas
matplotlib
scikit-learn
keras
tensorflow
twelvedata
python-dotenv
openpyxl
